+++ closedxml/Connections/IvyClosedXml/WorkbookRepository.cs
@@ -3,49 +3,70 @@ using System.Data;
 using ClosedXML.Excel;
 
 /// <summary>
-/// Repository class for managing Excel Workbooks 
-/// by creating, removing and storing current workbook on which operations are made
+/// Represents a repository for managing Excel workbook files using ClosedXML.
+/// This class handles operations such as adding, removing, and manipulating workbook files,
+/// including setting a current file for operations and saving data tables to worksheets.
 /// </summary>
 public class WorkbookRepository
 {
     private List<WorkbookFileInfo> excelFiles { get; set; }
     private WorkbookFileInfo? currentFile { get; set; }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="WorkbookRepository"/> class.
+    /// </summary>
     public WorkbookRepository()
     {
         excelFiles = new List<WorkbookFileInfo>();
     }
 
     /// <summary>
-    /// Set file on which operations are made.
+    /// Sets the current file on which operations are performed.
     /// </summary>
-    /// <param name="fileName"></param>
-    /// <returns></returns>
+    /// <param name="fileName">The name of the file to set as current.</param>
     public void SetCurrentFile(string fileName)
     {
         currentFile = GetFileByName(fileName);
     }
 
+    /// <summary>
+    /// Gets the current workbook file.
+    /// </summary>
+    /// <returns>The current <see cref="WorkbookFileInfo"/> instance.</returns>
     public WorkbookFileInfo GetCurrentFile()
     {
         return currentFile;
     }
 
+    /// <summary>
+    /// Gets the first table from the current worksheet as a DataTable.
+    /// If no table exists, returns a new DataTable with the name "FirstTable".
+    /// </summary>
+    /// <returns>A <see cref="DataTable"/> representing the first table in the worksheet.</returns>
     public DataTable GetCurrentTable()
     {
         var table = Worksheet.Tables.FirstOrDefault();
-
-        Console.WriteLine($"GetCurrentTable is {table}");
         if (table == null)
             return new DataTable() { TableName = "FirstTable" };
 
         return table.AsNativeDataTable();
     }
 
+    /// <summary>
+    /// Gets a read-only collection of all workbook files in the repository.
+    /// </summary>
+    /// <returns>A <see cref="ReadOnlyCollection{WorkbookFileInfo}"/> of workbook files.</returns>
     public ReadOnlyCollection<WorkbookFileInfo> GetFiles()
     {
         return new ReadOnlyCollection<WorkbookFileInfo>(excelFiles);
     }
 
+    /// <summary>
+    /// Adds a new workbook file to the repository.
+    /// </summary>
+    /// <param name="fileName">The name of the file to add.</param>
+    /// <exception cref="ArgumentException">Thrown if the file name already exists.</exception>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown if the file name is invalid.</exception>
     internal void AddNewFile(string fileName)
     {
         ValidateFileName(fileName);
@@ -56,6 +77,11 @@ public class WorkbookRepository
         excelFiles.Add(new WorkbookFileInfo(fileName));
     }
 
+    /// <summary>
+    /// Validates the provided file name for invalid characters and formats.
+    /// </summary>
+    /// <param name="fileName">The file name to validate.</param>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown if the file name is empty, contains invalid characters, or ends with invalid characters.</exception>
     private void ValidateFileName(string fileName)
     {
         if (string.IsNullOrWhiteSpace(fileName))
@@ -70,6 +96,11 @@ public class WorkbookRepository
             throw new ArgumentOutOfRangeException("File ends with invalid character!");
     }
 
+    /// <summary>
+    /// Removes a workbook file from the repository and disposes of its workbook.
+    /// If no files remain, sets the current file to null.
+    /// </summary>
+    /// <param name="fileName">The name of the file to remove.</param>
     internal void RemoveFile(string fileName)
     {
         if (ContainsFile(fileName))
@@ -85,16 +116,31 @@ public class WorkbookRepository
         }
     }
 
+    /// <summary>
+    /// Checks if a file with the specified name exists in the repository.
+    /// </summary>
+    /// <param name="fileName">The name of the file to check.</param>
+    /// <returns>True if the file exists; otherwise, false.</returns>
     private bool ContainsFile(string fileName)
     {
         return excelFiles.Any(f => f.FileName.Equals(fileName, StringComparison.InvariantCultureIgnoreCase));
     }
 
+    /// <summary>
+    /// Retrieves a workbook file by its name.
+    /// </summary>
+    /// <param name="fileName">The name of the file to retrieve.</param>
+    /// <returns>The <see cref="WorkbookFileInfo"/> instance matching the file name.</returns>
     private WorkbookFileInfo GetFileByName(string fileName)
     {
         return excelFiles.Single(f => f.FileName.Equals(fileName, StringComparison.InvariantCultureIgnoreCase));
     }
 
+    /// <summary>
+    /// Saves the provided DataTable to the current worksheet as a table named "FirstTable".
+    /// Removes any existing table with the same name before saving.
+    /// </summary>
+    /// <param name="table">The <see cref="DataTable"/> to save.</param>
     internal void Save(DataTable table)
     {
         TryRemoveExistingTable();
@@ -106,16 +152,20 @@ public class WorkbookRepository
         GetCurrentFile().Workbook.SaveAs(currentFile.FileName + ".xlsx");
     }
 
-
+    /// <summary>
+    /// Gets the first worksheet from the current workbook.
+    /// </summary>
     private IXLWorksheet Worksheet => currentFile.Workbook.Worksheets.FirstOrDefault();
     
+    /// <summary>
+    /// Attempts to remove the existing table named "FirstTable" from the worksheet.
+    /// Clears the worksheet if the table exists.
+    /// </summary>
     private void TryRemoveExistingTable()
     {
         try
         {
             var table = Worksheet.Tables.FirstOrDefault(f => f.Name == "FirstTable");
-
-            Console.WriteLine($"Table is {table}");
 
             if (table != null)
             {
@@ -124,9 +174,9 @@ public class WorkbookRepository
                 Worksheet.Clear(XLClearOptions.All);
             }
         }
-        catch (Exception e)
+        catch
         {
-            Console.WriteLine($"{e.Message}");
+            //ignored
         }
     }
 }
